CY-3435 get_all_results: omit the limit/offset clauses
In `_paginate`, make `get_all_results` just omit limit & offset.
Before, it was just doing nothing.

Also, the test was creating 1001 items, and then checking
that with `get_all_results` it returns 1000 (not all ie. 1001).
WHAT THE HELL.

Anyway, fixed the test, and also made it be 10 & 20 instead of 1001,
because inserting 1000 items via the storage managers takes on
the order of 10 seconds.